backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
chromadb
openai
langchain
speechrecognition
pyttsx3
